Add LtdAPI with cached auth tokens
The get_auth_token method will provide a cached token with a TTL of 10
minutes; which is within the validity of the tokens.

This will allow the client to make several API requests without needing
to get a new token each time.

Add methods for gettnig projects and organization resources, and return
those datasets in typed Pydantic objects.