Demonstrate ingesting an SSDC-MS from the web
This demonstrates an ingest pipeline for document metadata that starts
with LTD and downloads the metadata.json files for individual SSDC-MS
documents, and transforms that source metadata into domain objects.

Similar pipelines can be set up for other document types. This doesn't
include integrations with GitHub's API yet, though, which is needed to
get information about document issues and PRs.