Use PydanticRedisStorage for persistent storage
This change switches the ProjectRepository to now store domain models
for documentation projects in Redis, rather than in memory. This is
done so that implement asynchronous updates to the data (via GitHub
webhooks or periodic refreshes).

This is principally done through the PydanticRedisStorage class from
Safir (https://safir.lsst.io/user-guide/pydantic-redis.html). All domain
objects are now Pydantic objects rather than dataclasses to enable data
serialization to Redis.

Because the data is now requested asynchronously, there's a small change
to how the data is referenced in the Jinja templates, but otherwise the
update is straightforward.