Add ProjectService.ingest_github_project method
This method takes a github repo (based on the owner and repo name),
loads data from the GitHub API to populate a
SpherexGitHubSoftwareProject object, and upserts it into the redis-based
software project store.

Also added additional fields to the GitHubRepositoryModel to get
additional repository metadata such as the homepage URL, GitHub topics,
and description.