Add ProjectService.ingest_installed_github_repos
This method effectively upserts data for all github repositories for
which the GitHub app is installed in the configured GitHub organizations
(see the PORTAL_GITHUB_OWNERS configuration).

This method is called by the same bootstrap_from_api method that drives
the bulk ingestion of documents from the LTD service either on start up
or via the 15 minute arq cron task.